+++ ChangeLog.md
@@ -10,6 +10,7 @@
     - Rename ModalInfo global export to ModalInfo
     - Bump version requirements: Meteor to 2.13.2, pwix:i18n to 1.5, pwix:modal to 1.7
     - Define ModalInfo.i18n.namespace() (todo #4)
+    - Define classButton parameter (todo #5)
 
 ### 1.3.0