+++ ChangeLog.md
@@ -11,6 +11,7 @@
     - Bump version requirements: Meteor to 2.13.2, pwix:i18n to 1.5, pwix:modal to 1.7
     - Define ModalInfo.i18n.namespace() (todo #4)
     - Define classButton parameter (todo #5)
+    - Remove (useless) miDialog component
 
 ### 1.3.0